+++ src/Queue/Marshal/MarshalException.php
@@ -1,0 +1,8 @@
+<?php
+namespace Disque\Queue\Marshal;
+
+use Disque\Queue\QueueException;
+
+class MarshalException extends QueueException
+{
+}