+++ src/Queue/MarshalException.php
@@ -1,6 +1,0 @@
-<?php
-namespace Disque\Queue;
-
-class MarshalException extends QueueException
-{
-}